Improved piral-debug-utils
