+++ src/framework/piral-core/src/tools/codegen.ts
@@ -179,8 +179,8 @@ export function createRouteHandler(imports: Array<string>, exports: Array<string
   `);
 
   if (emulator) {
-    imports.push(`import { debugRouteFilter } from 'piral-debug-utils';`);
-    assignments.push('return debugRouteFilter(paths);');
+    imports.push(`import { useDebugRouteFilter } from 'piral-debug-utils';`);
+    assignments.push('return useDebugRouteFilter(paths);');
   } else {
     assignments.push('return paths;');
   }
