+++ src/utilities/piral-debug-utils/src/index.ts
@@ -1,4 +1,5 @@
 export * from './debug';
 export * from './emulator';
 export * from './types';
+export * from './useDebugRouteFilter';
 export * from './web';
