+++ src/utilities/piral-debug-utils/src/routeRefresh.ts
@@ -1,0 +1,17 @@
+export const debugRouteCache = {
+  active: 0,
+  paths: [],
+  refresh: undefined,
+};
+
+export function freezeRouteRefresh() {
+  debugRouteCache.active++;
+
+  return () => {
+    debugRouteCache.active--;
+
+    if (!debugRouteCache.active) {
+      debugRouteCache.refresh?.((s: number) => s + 1);
+    }
+  };
+}
